# go/model_error_model.go
/*
 * Gareth's Beer API
 *
 * A simple demonstration API exposing operations for working with Beers
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type ErrorModel struct {

	Code int32 `json:"code"`

	Message string `json:"message"`
}