# go/model_new_beer.go
/*
 * Gareth's Beer API
 *
 * A simple demonstration API exposing operations for working with Beers
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type NewBeer struct {

	Name string `json:"name"`

	Brewer string `json:"brewer,omitempty"`
}